fix other teachers not seeing an awarded badge
If an instructor awarded a badge to a student, the award wasn't seen
by other instructors. This lead to an error when that instructor
would attempt to give an award to the same student, since an award
already existed.

closes GRADE-2175

Test Plan
- Create a Badge in Roll Call.
- Have at least 2 instructors and 1 student.
- As Instructor A, award the Badge to the student.
- As Instructor B, verify that you can
  1) See that the Badge was awarded
  2) Un-award the Badge
  3) Re-award the Badge
- Verify that these actions propogate to the Instructor A as well.

Change-Id: I36de03afe4f4a3fa909ebe964b4701f6d9af126d
Reviewed-on: https://gerrit.instructure.com/192900
Tested-by: Jenkins
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Keith Garner <kgarner@instructure.com>
QA-Review: Gary Mei <gmei@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>