assume report csvs coming from canvas are UTF-8
When reports CSVs are snagged from canvas's CDN in production, the
Content-Type is correctly reported as text/csv, but there is no
associated charset set with it. Between ruby itself and httparty's
TextEncoder, the end result is strings from the CSV being encoded as
ASCII-8BIT/binary which causes some issues down the line when combined
with other data. This patchset adds a custom parser that assumes if the
content it was handed is ASCII-8BIT to force/tread it us UTF-8. This
parser will be passed to httpary to be used when downloadings CSV's from
canvas.

fixes TALLY-427

test plan:
 - Have a course with extended ascii characters and section names with
   extend ascii characters.
 - Mark students for attendance
 - As a canvas admin, run a 7 day report over the time you marked
   attendance
 - Ensure the report was correctly generated

Change-Id: Ide06c7dfbf38487f99b64a84da5530ca8d0ec512
Reviewed-on: https://gerrit.instructure.com/c/rollcall-attendance/+/227909
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Keith Garner <kgarner@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>