fix name highlighting in student list
In TALLY-427 a cross-site scripting vulnerability was fixed, but it
introduced a regression when the last cluster of characters were not
"word characters" and would display "undefined" after the name. This
patchset fixes that bug, but with the behavior that nothing will be
highlighted in those cases.

fixes TALLY-571

test plan:
 - Have a course set up with rollcall
 - Have three students in the source, one with a "plain" name, one who's
   name ends in parentheses, and one that just ends in other non-word
   characters. Examples: "Jane Smith", "Jonathon Guy (Jon)" and "Katie
   Money $$$"
 - Load rollcall, note that the plan name has the surname highlighted,
   and the other two names have no highlights but appear properly

Change-Id: I368606c3240bb2ab1257924d02f4647a9479a3cb
Reviewed-on: https://gerrit.instructure.com/c/rollcall-attendance/+/229322
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Gary Mei <gmei@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>