refactor assignment retrieval strategy
We're still seeing a large number of failures (20k an hour) in splunk,
especially when it seems a whole section is being marked. The default
retry for lock aquisition is 200ms which is too fast. We'll bump the
time between retries to 1 second, with the default three retries, we'll
wait up to 3 seconds. Addtionally, we don't want grade updates to fail,
so we'll attempt this whole process up to three times before we give up.

This makes our max delay up to 9 seconds, and if it doesn't work in 9
seconds we have much larger problems.

Also, since this is a background job, we can endure the maximum delay.

Change-Id: Ibbabff77c1d24f9024a7a596ceec6ebfea9ddfa3
Reviewed-on: https://gerrit.instructure.com/c/rollcall-attendance/+/340503
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Hilario <luis.hilario@instructure.com>
QA-Review: Luis Hilario <luis.hilario@instructure.com>
Product-Review: Luis Hilario <luis.hilario@instructure.com>